% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{copulaGarch.sim}
\alias{copulaGarch.sim}
\title{Simulate from copula-garch models given start values or continue from given data}
\usage{
copulaGarch.sim(n, type, param, sig0, eps0, returns0, Y)
}
\description{
@param n Number of simulated data points
 @param type type of copula. One of "gaussian" or "t"
 @param param List of copula-garch model parameters.
 @param sig0 Vector of start values of sigma for each stock. If not provided, will continue from Y.
 @param eps0 Vector of start values of residual for each stock. If not provided, will continue from Y.
 @param returns0 Vector of start values of returns for each stock. If not provided, will continue from Y.
 @param Y Data to continue with. Either Y or all of sig0, eps0 and returns0 need to provided
 @return Simulated copula-garch data.
 @export
}
