% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{dgCopula}
\alias{dgCopula}
\title{Compute dendity of Gaussian Copula}
\usage{
dgCopula(U, sigma, log = FALSE)
}
\arguments{
\item{sigma}{Covariate of gaussian copula}

\item{log}{Logical variable. If true, return log density}

\item{Input}{Variable. Distribution is U(0,1)}
}
\description{
Compute dendity of Gaussian Copula
}
