% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{dtCopula}
\alias{dtCopula}
\title{Compute dendity of t Copula}
\usage{
dtCopula(u, sigma, df, log = FALSE)
}
\arguments{
\item{sigma}{Covariate of t copula}

\item{df}{Degrees of freedom of t copula}

\item{log}{Logical variable. If true, return log density}

\item{Input}{Variable. Distribution is U(0,1)}
}
\description{
Compute dendity of t Copula
}
