% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{fitCopula.GT}
\alias{fitCopula.GT}
\title{Fit a gaussian or t copula and return copula parameters using maximum likelihood method}
\usage{
fitCopula.GT(U, type)
}
\arguments{
\item{U}{nxd matrix of observations with distribution U~(0,1)}

\item{type}{Type of copula to fit. One of "gaussian" and "t"}
}
\value{
Returns fitted copula variables. Sigma for both gaussian and t copula and df for t copula only
}
\description{
Fit a gaussian or t copula and return copula parameters using maximum likelihood method
}
\details{
Using \eqn{optim} to optimize loglikelihood function. The method used in optimize is "BFGS".
}
