% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{fitgarchCopula}
\alias{fitgarchCopula}
\title{Fit a Corpula-Garch(1,1) using maximum likelihood method}
\usage{
fitgarchCopula(Y, type)
}
\arguments{
\item{Y}{a univariate data object.}

\item{type}{Type of Copula-Garch model, one of "gaussian" and "t"}
}
\value{
A list of Copula-Garch(1,1) parameters
}
\description{
Fit a Corpula-Garch(1,1) using maximum likelihood method
}
\details{
The garch model used in the Copula-Garch is \eqn{"sGarch"} with a standard t residual distribution.
The maximum-likeihood is optimized using \eqn{optim} function, with parameters set to default.
}
