% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{garch.loglik}
\alias{garch.loglik}
\title{GARCH loglikelihood function.}
\usage{
garch.loglik(omega, alpha, beta, df, eps, sig2)
}
\arguments{
\item{omega, alpha, beta}{GARCH parameters (scalars; see \code{\link{garch.sim}}).}

\item{eps}{Vector of GARCH observations \code{eps1, ..., epsN}.}
}
\value{
The GARCH log-likelihood.
}
\description{
GARCH loglikelihood function.
}
\details{
Validates parameters by checking that \code{all(sig2 > 0) == TRUE}.
}
