% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{Garchcop.LogLik}
\alias{Garchcop.LogLik}
\title{Log-likelihood function of Copula-Garch(1,1), with Gaussian Copula or t Copula.}
\usage{
Garchcop.LogLik(alpha, beta, omega, mu, shape, rho, df, type, Y)
}
\arguments{
\item{alpha}{A vector contains alpha values for Garch model for each stock.}

\item{beta}{A vector contains beta valus for Garch model for each stock.}

\item{omega}{A vector contains omega values for Garch model for each stock.}

\item{mu}{A vector contains mu values for each Garch model for each stock.}

\item{shape}{A vector contains the degrees of freedom for Garch residual distribution for each stock.}

\item{rho}{A vector contains the upper triangular values of correlation matrix for the copula.}

\item{df}{A vector contains the degrees of freedom for the copula. Only needed for t-Copula.}

\item{type}{Type of copula. "gaussian" or "t".}

\item{Y}{A matrix, vector or xts object of returns.}
}
\value{
Loglikelihood of Copula-Garch
}
\description{
Log-likelihood function of Copula-Garch(1,1), with Gaussian Copula or t Copula.
}
