% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{Garchgc.sim}
\alias{Garchgc.sim}
\title{A Garch GC simulate function that takes fitted garch models}
\usage{
Garchgc.sim(n, model, var, D)
}
\arguments{
\item{n}{Number of simulated data points}

\item{model}{List of fitted garch models}

\item{D}{number of stocks in total}

\item{List}{contains the sectir-wised covariance matrix}
}
\value{
Simulated data
}
\description{
A Garch GC simulate function that takes fitted garch models
}
