% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{gCopula.LogLik}
\alias{gCopula.LogLik}
\title{Gaussian-Copula Log-Likelihood}
\usage{
gCopula.LogLik(sigma, U)
}
\arguments{
\item{sigma}{Upper triangular of sigma for multi-variate gaussian model}

\item{U}{nxd matrix of observations with distribution U~(0,1)}
}
\value{
Log-likelihood of Gaussian Copula
}
\description{
Gaussian-Copula Log-Likelihood
}
