% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{getreturns}
\alias{getreturns}
\title{Compute returns from given data}
\usage{
getreturns(S, type)
}
\arguments{
\item{S}{A nxd matrix, vector or \eqn{xts} object input into function to calculate returns from.}

\item{type}{Type of calculated return, one of "absolute", "relative" and "log".}
}
\value{
A matrix, vector or \eqn{xts} object of returns.
}
\description{
Compute returns from given data
}
\details{
Absolute returns: \eqn{y){ti} = s_{ti} - s_{t-1,i}}
Relative returns: \eqn{y_s_{ti} - (s_{ti} - s_{t-1,i})/s_{t-1,i}}
Log returns: \eqn{y_{ti} = log(s_{ti}) - log(s_{t-1,i})}
}
