% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{hier.A}
\alias{hier.A}
\title{Use in the maximization step of hierarchical modeling for updating the value of Sigma}
\usage{
hier.A(S)
}
\arguments{
\item{S}{list denotes the distribution parameter of S}
}
\value{
A
}
\description{
Use in the maximization step of hierarchical modeling for updating the value of Sigma
}
