% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{hier.B}
\alias{hier.B}
\title{Used in the maximization step of hierarchical modelling for updating the value of Psi}
\usage{
hier.B(S, Z, group)
}
\arguments{
\item{S}{List denotes the distribution of S}

\item{Z}{Matrix denote the normalized residual of of each stock for each day}

\item{group}{Length-\code{N} vector of integers between 1 and \code{K} of class labels.}
}
\value{
B for each sector
}
\description{
Used in the maximization step of hierarchical modelling for updating the value of Psi
}
