% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{hier.logLik}
\alias{hier.logLik}
\title{Compute the Log Likelihood of hierarchical multivariate normal distribution}
\usage{
hier.logLik(S, Z, Sigma, Psi, group)
}
\arguments{
\item{S}{Matrix denote the mean of sector K for each day}

\item{Z}{Matrix denote the normalized residual of of each stock for each day}

\item{Sigma}{\code{K x K} intergroup variance matrix.}

\item{Psi}{List of \code{K} intragroup variance matrices.}

\item{group}{Length-\code{N} vector of integers between 1 and \code{K} of class labels.}
}
\value{
Loglikelihood of sector-level hierarchical modelling
}
\description{
Compute the Log Likelihood of hierarchical multivariate normal distribution
}
