% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{hier.var}
\alias{hier.var}
\title{Intergroup and intragroup variance matrix.}
\usage{
hier.var(Sigma, Psi, group)
}
\arguments{
\item{Sigma}{\code{K x K} intergroup variance matrix.}

\item{Psi}{List of \code{K} intragroup variance matrices.}

\item{group}{Length-\code{N} vector of integers between 1 and \code{K} of class labels.}
}
\value{
Variance matrix of size \code{(N+K) x (N+K)}.
}
\description{
Intergroup and intragroup variance matrix.
}
\details{
Returns the variance of \code{(S, V)}, where
\preformatted{
  S = (S_1, ..., S_K) ~ N(0, Sigma)
  V_k = (V(k)_1, ..., V(K)_NK) | S ~ind N(S_k, Psi_k),
}
and \code{N = N1 + ... + NK}.
}
