% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{pgCopula}
\alias{pgCopula}
\title{Conputes distribution function of Gaussian Copula}
\usage{
pgCopula(U, sigma)
}
\arguments{
\item{U}{Input Variables. Distribution is U(0,1)}

\item{sigma}{Covariance of gaussian copula}
}
\value{
Distribution function of input
}
\description{
Conputes distribution function of Gaussian Copula
}
