% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{ptCopula}
\alias{ptCopula}
\title{Conputes distribution function of t Copula}
\usage{
ptCopula(U, sigma, df)
}
\arguments{
\item{U}{Input Variables. Distribution is U(0,1)}

\item{sigma}{Covariance of t copula}

\item{df}{Degrees of freedom of t copula}
}
\value{
Distribution function of input
}
\description{
Conputes distribution function of t Copula
}
