% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{rgCopula}
\alias{rgCopula}
\title{Generate random Gaussian Copula variates}
\usage{
rgCopula(n, sigma)
}
\arguments{
\item{n}{Number of observations generated}

\item{sigma}{Sigma value for the multivariate gaussian model}
}
\value{
Generated random gaussian-copula variate
}
\description{
Generate random Gaussian Copula variates
}
