% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{rtCopula}
\alias{rtCopula}
\title{Generate random t-Copula variates}
\usage{
rtCopula(n, sigma, df)
}
\arguments{
\item{n}{Number of observations generated}

\item{sigma}{Sigma value for the multivariate t model}

\item{Degrees}{of freedom for the t-copula}
}
\value{
Generated random t-copula variate
}
\description{
Generate random t-Copula variates
}
