% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{S.expect}
\alias{S.expect}
\title{Computes the conditional expectation of stock price given normalized residual}
\usage{
S.expect(Sigma, Psi, Z, group)
}
\arguments{
\item{Sigma}{\code{K x K} intergroup variance matrix.}

\item{Psi}{List of \code{K} intragroup variance matrices.}

\item{Z}{Matrix denote the normalized residual of of each stock for each day}

\item{group}{Length-\code{N} vector of integers between 1 and \code{K} of class labels.}
}
\description{
Computes the conditional expectation of stock price given normalized residual
}
