% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{tCopula.Loglik}
\alias{tCopula.Loglik}
\title{t-Copula Log-Likelihood}
\usage{
tCopula.Loglik(sigma, df, U)
}
\arguments{
\item{sigma}{Upper triangular of sigma for multi-variate t model}

\item{df}{Degrees of freedom for multi-variate t model}

\item{U}{nxd matrix of observations with distribution U~(0,1)}
}
\value{
Log-likelihood of t-Copula
}
\description{
t-Copula Log-Likelihood
}
